feat: refactor handlers and web related tooling into web subdir
Closes #30
