+++ src/db/mock.rs
@@ -1,7 +1,3 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, you can obtain one at https://mozilla.org/MPL/2.0/.
-
 //! Mock db implementation with methods stubbed to return default values.
 
 use futures::future;
