+++ src/web/mod.rs
@@ -1,0 +1,2 @@
+pub mod auth;
+pub mod handlers;
